+++ libs/form-builder/src/components/CheckRender.tsx
@@ -2,7 +2,7 @@ import { ComponentProps } from 'react';
 import { useFormContext } from 'react-hook-form';
 import { Buildable, CheckGate } from '@daohaus/ui';
 
-import { FormBuilderFactory } from './FormBuilderFactory';
+import { FormBuilderFactory } from '@daohaus/form-builder-base';
 import { FieldLego } from '../types';
 
 type CheckRenderProps = Omit<
