+++ libs/form-builder/src/components/index.ts
@@ -2,5 +2,3 @@ export * from './CoreFieldLookup';
 export * from './FieldSpacer';
 export * from './ToWeiInput';
 export * from './formFooter';
-export { useFormBuilder } from '@daohaus/form-builder-base';
-export { FormBuilderFactory } from '@daohaus/form-builder-base';
