+++ libs/form-builder/src/index.ts
@@ -1,3 +1,5 @@
 export * from './components';
 export * from './types';
 export * from './FormBuilder';
+export { useFormBuilder } from '@daohaus/form-builder-base';
+export { FormBuilderFactory } from '@daohaus/form-builder-base';
